requirements.txt
argon2-cffi==23.1.0
Flask==2.2.2
Flask_SocketIO==5.3.6
qrcode==7.4.2
